feat: add custom filter search fields
