+++ web/src/locales/id/data.json
@@ -308,6 +308,8 @@
     "Edit LDAP": "Mengedit LDAP",
     "Enable SSL": "Aktifkan SSL",
     "Enable SSL - Tooltip": "Apakah untuk mengaktifkan SSL?",
+    "Filter fields": "Filter fields",
+    "Filter fields - Tooltip": "Filter fields - Tooltip",
     "Group ID": "ID grup",
     "Last Sync": "Terakhir Sinkronisasi",
     "Search Filter": "Search Filter",
